Add fixtures for DS.FixtureAdapter
I don’t like that this is in the model files, rather than in a fixture
directory.  It might make sense to try to rig this up using an
initializer, possibly the same one used for Pretender, since they’re
working together.  They could be toggled depending on the environment
in one easy step that way, also.